refactor: Extract JobPool, Request from ConcurrentRequestExecutor to facilitate testing
